core/requirements.txt
Django==4.1.1
django-s3-sqlite==0.0.5
boto3==1.24.66
factory-boy==3.2.1
awslambdaric==2.0.4
serverless-wsgi==3.0.1
werkzeug==2.2.2
django-storages==1.13.1
django-bootstrap-v5==1.0.11
django-extensions==3.2.1
pytest-xdist==3.0.2
pytest-django==4.5.2
Faker==15.1.3
pycognito==2022.8.0
watchtower==1.0.6
django-selenium-login==2.0.0
pytest-selenium==4.0.0
pytest-factoryboy==2.5.0
